incluido stdlib no modulo de listas ligadas; modificação dos campos de peso e custo do camiao (agora são double, em vez de float); criada função para ler unsigned int e para ler os doubles; essa função prevê comportamentos como escrever letras em vez de um double e nesses casos faz return de NaN

+++ dados.c
@@ -22,7 +22,7 @@ void camiao_dump( void* camiao ){
     printf("{id=%3d, matricula=\"%s\", consumo=%f}\n", thisCamiaoPt->id, thisCamiaoPt->matricula, thisCamiaoPt->custo );
 }
 
-CamiaoPt camiao_novo( unsigned int id, char *matricula, float custo, float peso ){
+CamiaoPt camiao_novo( unsigned int id, char *matricula, double custo, double peso ){
     CamiaoPt novoCamiaoPt = NULL;
     if( (novoCamiaoPt = malloc( sizeof(Camiao)) ) == NULL ) return NULL;
